// src/engines/frequency-city/frequencyCity.types.ts
export type FrequencyCityConfig = {
  pulseSpeed: number;
  energyResponse: number;
  buildingCount: number;
  glowIntensity: number;
  spaceScale: number;
  colorSaturation: number;
  sparkleDensity: number;
  warmth: number;
  magentaAccent: string;
  cyanAccent: string;
  violetAccent: string;
  showTitle: boolean;
};